+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
   - `DefaultSirenElementVisitor` for visitors that don't need to visit every type of element
   - `LinkFinder` for searching an entity's links
   - `ActionFiller` for populating an action's fields
+  - `CompositeVisitor` for composing multiple visitors
 - Support for validating fields on action submission ([#16](https://github.com/siren-js/client/issues/16))
 - Default JSON serializer ([#27](https://github.com/siren-js/client/issues/27))
 - Advanced `submit` usage examples the JSDocs